Implemented issue #23
Implemented a pass collecting data about callsites number for each function. This information is collected for input
independent functions only.
Additionally the pass collects functions that are called from loops. If a function is called in a loop then the pass
does not collect information out callsites number for it.

Removed unused passes.