Implemented first part of issue #45 - dg integration with oh.
Modified dg CMakefiles to install all necessary headers and libraries to system directories.
Compiled dg with OH as the first step of integration.
In OH created a wrapper for dg slicer which task is to run dg slicer and record instructions/basic blocks contained in a
slice. Slicer wrapper gets function for which slicing has to be done and unique assertion name which is used as slicing
criterion.