Implemented issue #49
Before processing oh path check if an argument of a function is used in it, if so do not process the path. Skip data
dependent uses as those will be skipped later in extraction phase. Skip uses in deterministic blocks as those will be
removed later when extracting oh path function