Modified insertHashBuilder as discussed in issue #68.
Added a check for instruction/value type before calling insertHashBuilder. If the check doesn't succeed instruction is
added to the set of "non-hashable" instructions and insertHashBuilder is not invoked. In insertHashBuilder added an
assertion for the validity of a value passed to it, i.e. integer or float type.
Removed the error message mentioned in the issue.